feat: implement intercepting window.fetch
